changes from int to floats

+++ Elevator.py
@@ -4,13 +4,13 @@ class Elevator:
     def __init__(self, id: float, speed: float, minFloor: float, maxFloor: float, closeTime: float, openTime: float, startTime: float,
                  stopTime: float) -> None:
         self.id = id
-        self.speed = int(speed)
-        self.minFloor = int(minFloor)
-        self.maxFloor = int(maxFloor)
-        self.closeTime = int(closeTime)
-        self.openTime = int(openTime)
-        self.startTime = int(startTime)
-        self.stopTime = int(stopTime)
+        self.speed = float(speed)
+        self.minFloor = float(minFloor)
+        self.maxFloor = float(maxFloor)
+        self.closeTime = float(closeTime)
+        self.openTime = float(openTime)
+        self.startTime = float(startTime)
+        self.stopTime = float(stopTime)
         self.calls = []
